固定スライドのスクロールコントロール用コンポーネントを切り出し
