+++ components/gimmik/organisms/FixedPresentation.module.scss
@@ -50,33 +50,6 @@
     height: 100%;
   }
 }
-.fixed-presentation-controller {
-  position: relative;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 100vw;
-  height: 100%;
-  background-color: red;
-  &--content {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    padding: 1px 0;
-    box-sizing: content-box;
-    background-color: blue;
-    opacity: 0.3;
-  }
-}
-
-@media (max-width: #{var.$breakpoint}px) {
-  .pzl-font-grid {
-    &__nav {
-      button {
-        right: 20px;
-      }
-    }
-  }
-}
 
 @media (min-width: #{var.$breakpoint-pc + 1}px) {
   .fixed-presentation {
