+++ components/gimmik/organisms/FixedSliderController.module.scss
@@ -1,0 +1,23 @@
+/*===================================*/
+// FixedSlider STYLE
+// ロゴエリアコンポーネントのスタイル
+/*===================================*/
+@use "../../../styles/var.module";
+
+.fixed-presentation-controller {
+  position: relative;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 100vw;
+  height: 100%;
+  background-color: red;
+  &--content {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    padding: 1px 0;
+    box-sizing: content-box;
+    background-color: blue;
+    opacity: 0.3;
+  }
+}
